chore: Divide the code related to ProductHistory
